removing ssao as it is buggy and annoyinng

+++ OpenResearchEngine/Render/Resources/RenderTextures.h
@@ -19,8 +19,8 @@ public:
     std::vector<CD3DX12_CPU_DESCRIPTOR_HANDLE> GetCpuSrvHandles() { return cpuSrvHandles; }
     std::vector<CD3DX12_GPU_DESCRIPTOR_HANDLE> GetGpuSrvHandles() { return gpuSrvHandles; }
 
-    int GetNumCpuSrvHandles() { return cpuSrvHandles.size(); }
-    int GetNumGpuSrvHandles() { return gpuSrvHandles.size(); }
+    size_t GetNumCpuSrvHandles() { return cpuSrvHandles.size(); }
+    size_t GetNumGpuSrvHandles() { return gpuSrvHandles.size(); }
 
     D3D12_CPU_DESCRIPTOR_HANDLE GetStartCpuSrv() const { return startCpuSrv; }
     D3D12_GPU_DESCRIPTOR_HANDLE GetStartGpuSrv() const { return startGpuSrv; }
